fix(site): keep leva, the frame counter and the space toggle out of production
The dev overlays were guarded with `import.meta.env.DEV`, which is a
build-time constant, so the branches did fold away. The imports did not:
leva builds its stitches theme at module scope, and a production build
carried its whole panel UI — theme, CSS-in-JS runtime and all.

Split the look into a module production actually needs (`live`, read every
frame) and a dev module that writes to it, and reach the dev modules only
through `import()` from a branch that folds to `null`. Rolldown then never
emits the chunk.

Measured on `vite build`:

  bundle   3,301.75 kB -> 3,117.62 kB   (gzip 975.85 -> 913.60)
  emitted dev chunks: none
  `titleBar`, `stitches`, `__look`: absent

The remaining `leva` substring matches are inside our own `keyElevation`.

Also guard the nullable `renderPipeline`: r3f alpha.3 types it `T | null`,
since the callback also runs for teardown.
